+++ src/sirabus/__init__.py
@@ -1,4 +1,3 @@
-import abc
 from abc import ABC, abstractmethod
 from typing import Generic, Type, TypeVar, get_args, Optional
