+++ src/sirabus/message_pump.py
@@ -66,7 +66,9 @@ class MessagePump:
                         ]
                     )
                 )
-                logging.info(f"Processed message with headers: {headers} and body: {body}")
+                logging.info(
+                    f"Processed message with headers: {headers} and body: {body}"
+                )
             else:
                 time.sleep(0.1)
